requirements.txt
Flask
python-dotenv
cloudinary
requests
